+++ TR5Main/Objects/TR1/Entity/tr1_wolf.cpp
@@ -70,9 +70,9 @@ void WolfControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short head = 0;
-	short angle = 0;
-	short tilt = 0;
+	float head = 0;
+	float angle = 0;
+	float tilt = 0;
 
 	if (item->HitPoints <= 0)
 	{
